# src/docs/swagger.yaml
openapi: 3.0.3
info:
  title: Rain
  description: Rain API's Documentation
  version: 1.0.0
servers:
  - url: http://localhost:3000
    description: Local server
...
security:
  - bearerAuth: []
paths:
  /auth/register-with-walletAddress:
    post:
      tags:
        - Auth
      summary: Register user with wallet address
      description: Registers a new user by verifying their wallet address using a signed message.
      operationId: registerWithWalletAddress
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                sign:
                  type: string
                  description: The signed message to verify the user's wallet address
                  example: "0x..."
                name:
                  type: string
                  description: The name of the user
                  example: "John Doe"
                bio:
                  type: string
                  description: The bio of the user (optional)
                  example: "Web3 enthusiast"
                object:
                  type: object
                  description: The object containing the message to be signed
                  properties:
                    address:
                      type: string
                      description: The wallet address of the user
                      example: "0x1234567890abcdef1234567890abcdef12345678"
                    message:
                      type: string
                      description: The message to be signed by the user
                      example: "Sign this message to verify ownership of the wallet."
      responses:
        200:
          description: User registered successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: boolean
                    example: true
                  msg:
                    type: string
                    example: "User Registered"
                  token:
                    type: string
                    example: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJpZCI6IjYwYjZkNzI4NmNhZTYzZTJiODMzNjEwZGVkYTA1ZTUzYTciLCJpYXQiOiJhbG9n..."
                  user:
                    type: object
                    properties:
                      _id:
                        type: string
                        example: "60b6c72b7c71bf76c5d4be65"
                      name:
                        type: string
                        example: "John Doe"
                      walletAddress:
                        type: string
                        example: "0x1234567890abcdef1234567890abcdef12345678"
        400:
          description: Bad request, missing required fields
          content:
            application/json:
              schema:
                type: object
                properties:
                  msg:
                    type: string
                    example: "All inputs are required"
        401:
          description: Unauthorized, signature verification failed or wallet address not correct
          content:
            application/json:
              schema:
                type: object
                properties:
                  msg:
                    type: string
                    example: "Sign not verified"
        409:
          description: Conflict, user already exists
          content:
            application/json:
              schema:
                type: object
                properties:
                  msg:
                    type: string
                    example: "User already exists"
        500:
          description: Internal Server Error, something went wrong on the server
          content:
            application/json:
              schema:
                type: object
                properties:
                  msg:
                    type: string
                    example: "Error during registration: <error message>"
  /auth/login-with-walletAddress:
    post:
      tags:
        - Auth
      summary: User login using wallet address and signed message
      description: Log in a user by verifying the signature of their wallet address.
      operationId: loginWithWalletAddress
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                walletAddress:
                  type: string
                  description: The wallet address of the user.
                  example: "0xF2Adc1FAC6629B567cF60154C3FEA4e498E58C91"
                sign:
                  type: string
                  description: The signed message to verify the user's wallet address.You can generate this signed message using a tool like [codesandbox.io](https://codesandbox.io/p/sandbox/react-eth-metamask-signatures-ibuxj?file=%2Fsrc%2FApp.js).
                  example: "0x..."
                object:
                  type: object
                  description: The object containing the message to be signed and wallet address.
                  properties:
                    address:
                      type: string
                      description: The wallet address of the user
                      example: "0x1234567890abcdef1234567890abcdef12345678"
                    message:
                      type: string
                      description: The message to be signed by the user
                      example: "Sign this message to verify ownership of the wallet."
      responses:
        200:
          $ref: "#/components/responses/LoginSuccess"
        400:
          description: Bad request, missing required fields
          content:
            application/json:
              schema:
                type: object
                properties:
                  msg:
                    type: string
                    example: "Sign and wallet address are required"
        404:
          $ref: "#/components/responses/UserNotFoundError"
        401:
          description: Unauthorized, signature verification failed
          content:
            application/json:
              schema:
                type: object
                properties:
                  msg:
                    type: string
                    example: "Sign not verified"
        500:
          description: Internal server error, something went wrong on the server
          content:
            application/json:
              schema:
                type: object
                properties:
                  msg:
                    type: string
                    example: "Error during login: <error message>"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  responses:
...
    LoginSuccess:
      description: Successful user login with wallet address
      content:
        application/json:
          schema:
            type: object
            properties:
              status:
                type: boolean
                example: true
              msg:
                type: string
                example: "User Logged In"
              token:
                type: string
                example: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
              user:
                type: object
                properties:
                  _id:
                    type: string
                    example: "60b6c72b7c71bf76c5d4be65"
                  name:
                    type: string
                    example: "John Doe"
                  walletAddress:
                    type: string
                    example: "0x1234567890abcdef1234567890abcdef12345678"
    UserNotFoundError:
      summary: User not found
      type: object
      properties:
        statusCode:
          type: integer
          example: 404
        message:
          type: string
          example: User not found
        details:
          example: null
...